# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Старт API",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/max": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Auth Max",
        "operationId": "auth_max_auth_max_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaxAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Login",
        "operationId": "admin_login_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/portfolio": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get My Portfolio",
        "operationId": "get_my_portfolio_me_portfolio_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioItem"
                  },
                  "type": "array",
                  "title": "Response Get My Portfolio Me Portfolio Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/portfolio/link": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get My Portfolio Link",
        "operationId": "get_my_portfolio_link_me_portfolio_link_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioLinkState"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "portfolio"
        ],
        "summary": "Regenerate My Portfolio Link",
        "operationId": "regenerate_my_portfolio_link_me_portfolio_link_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioLinkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Create My Portfolio Link",
        "description": "Turns the student's confirmed experience into a shareable public page.\n\nDeliberately explicit: nothing is public until the student asks for a\nlink. Calling it again returns the same slug.",
        "operationId": "create_my_portfolio_link_me_portfolio_link_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioLinkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Delete My Portfolio Link",
        "operationId": "delete_my_portfolio_link_me_portfolio_link_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/p/{slug}": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Public Portfolio",
        "description": "Public portfolio — no auth, no contact details, recruiter-friendly.",
        "operationId": "get_public_portfolio_p__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/bot/webhook": {
      "post": {
        "tags": [
          "bot"
        ],
        "summary": "Bot Webhook",
        "operationId": "bot_webhook_bot_webhook_post",
        "parameters": [
          {
            "name": "x-max-bot-api-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Bot-Api-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Bot Webhook Bot Webhook Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminLoginRequest": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "AdminLoginRequest"
      },
...
      "ConfirmationRead": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution"
          },
          "confirmed_by": {
            "type": "string",
            "title": "Confirmed By"
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "contribution",
          "confirmed_by",
          "confirmed_at"
        ],
        "title": "ConfirmationRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MaxAuthRequest": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Init Data"
          }
        },
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "MaxAuthRequest"
      },
...
      "OrganizationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "type",
          "verified"
        ],
        "title": "OrganizationRead"
      },
      "PortfolioItem": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectListItem"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectResultRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmation": {
            "$ref": "#/components/schemas/ConfirmationRead"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result",
          "confirmation"
        ],
        "title": "PortfolioItem"
      },
      "PortfolioLinkResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "slug"
        ],
        "title": "PortfolioLinkResponse"
      },
      "PortfolioLinkState": {
        "properties": {
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "slug"
        ],
        "title": "PortfolioLinkState"
      },
      "PortfolioPublic": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "preferred_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Role"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillPublic"
            },
            "type": "array",
            "title": "Skills"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/PortfolioPublicItem"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioPublic",
        "description": "Everything a recruiter/teacher sees on /p/<slug>. No contacts."
      },
      "PortfolioPublicItem": {
        "properties": {
          "project_title": {
            "type": "string",
            "title": "Project Title"
          },
          "organization": {
            "type": "string",
            "title": "Organization"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectResultRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "project_title",
          "organization",
          "role",
          "contribution",
          "result",
          "confirmed_at"
        ],
        "title": "PortfolioPublicItem"
      },
...
      "ProjectDifficulty": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced"
        ],
        "title": "ProjectDifficulty"
      },
...
      "ProjectFormat": {
        "type": "string",
        "enum": [
          "online",
          "offline",
          "hybrid"
        ],
        "title": "ProjectFormat"
      },
      "ProjectListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProjectDifficulty"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "format": {
            "$ref": "#/components/schemas/ProjectFormat"
          },
          "deadline": {
            "type": "string",
            "title": "Deadline"
          },
          "participant_limit": {
            "type": "integer",
            "title": "Participant Limit"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty",
          "status",
          "format",
          "deadline",
          "participant_limit",
          "organization"
        ],
        "title": "ProjectListItem"
      },
...
      "ProjectResultRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "result_url",
          "completed_at"
        ],
        "title": "ProjectResultRead"
      },
...
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "draft",
          "open",
          "in_progress",
          "completed"
        ],
        "title": "ProjectStatus"
      },
...
      "SkillPublic": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "level": {
            "type": "string",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "level"
        ],
        "title": "SkillPublic"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}